Mbnavadjust, macros, format MBF_MBARIROV (165)

+++ src/mbnavadjust/mbnavadjust_util.c
@@ -572,14 +572,14 @@ static XmString StringToXmString(char *str) {
 		 * Create the segment. Text and tag first.
 		 */
 		if (textLen) {
-			if (textBufLen <= (textLen * sizeof(wchar_t))) {
+			if (textBufLen <= (textLen * (int)sizeof(wchar_t))) {
 				textBufLen = (textLen + 1) * sizeof(wchar_t);
 				textBuf = (char *)XtRealloc(textBuf, textBufLen);
 			}
 			copyWcsToMbs(textBuf, text, textLen, True);
 
 			if (tagLen) {
-				if (tagBufLen <= (tagLen * sizeof(wchar_t))) {
+				if (tagBufLen <= (tagLen * (int)sizeof(wchar_t))) {
 					tagBufLen = (tagLen + 1) * sizeof(wchar_t);
 					tagBuf = (char *)XtRealloc(tagBuf, tagBufLen);
 				}
@@ -1097,7 +1097,7 @@ void BX_MENU_POST(Widget p, XtPointer mw, XEvent *ev, Boolean *dispatch) {
 
 	Arg args[2];
 	int argcnt = 0;
-	int button;
+	unsigned int button;
 	XtSetArg(args[argcnt], XmNwhichButton, &button);
 	argcnt++;
 	XtGetValues(m, args, argcnt);
@@ -1667,7 +1667,7 @@ LFUNC(atoui, unsigned int, (char *p, unsigned int l, unsigned int *ui_return));
 #endif
 
 static unsigned int atoui(char *p, unsigned int l, unsigned int *ui_return) {
-	int i;
+	unsigned int i;
 
 	int n = 0;
 	for (i = 0; i < l; i++)
@@ -2443,7 +2443,7 @@ static void SetImagePixels(
 	char *dst;
 	int nbytes;
 	unsigned int *iptr;
-	int x, y, i;
+	unsigned int x, y, i;
 
 	iptr = pixelindex;
 	if (image->depth == 1) {
@@ -2508,7 +2508,7 @@ static void SetImagePixels32(
 	unsigned char *addr;
 	unsigned int *paddr;
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	iptr = pixelindex;
 #ifndef WORD64
@@ -2549,7 +2549,7 @@ static void SetImagePixels16(
     XImage *image, unsigned int width, unsigned int height, unsigned int *pixelindex, Pixel *pixels) {
 	unsigned char *addr;
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	iptr = pixelindex;
 	if (image->byte_order == MSBFirst)
@@ -2576,7 +2576,7 @@ static void SetImagePixels8(
     XImage *image, unsigned int width, unsigned int height,
     unsigned int *pixelindex, Pixel *pixels) {
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	iptr = pixelindex;
 	for (y = 0; y < height; y++)
@@ -2594,7 +2594,7 @@ static void SetImagePixels1(
 	unsigned char bit;
 	int xoff, yoff;
 	unsigned int *iptr;
-	int x, y;
+	unsigned int x, y;
 
 	if (image->byte_order != image->bitmap_bit_order)
 		SetImagePixels(image, width, height, pixelindex, pixels);
